@@ -14,8 +14,8 @@
     "moment": "^2.29.4",
     "prop-types": "^15.7.2",
     "query-string": "^7.1.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-error-boundary": "^4.0.11",
     "react-helmet": "^6.1.0",
     "react-outside-click-handler": "^1.3.0",
@@ -62,8 +62,8 @@
     "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
     "@testing-library/cypress": "^10.0.1",
     "@testing-library/jest-dom": "^6.1.5",
-    "@testing-library/react": "^12.1.5",
-    "@testing-library/user-event": "^13.5.0",
+    "@testing-library/react": "^14.1.2",
+    "@testing-library/user-event": "^14.5.1",
     "axios-mock-adapter": "^1.22.0",
     "babel-eslint": "^10.1.0",
     "cypress": "^13.6.1",
